+++ src/pages/domain.tsx
@@ -8,6 +8,7 @@ import agent from '../api/agent';
 import { ILookupTable, ILookupTableLayout, PageProps, IDNSRecordDetails } from "../interfaces";
 import { IPAddressGeo, MyIpAddressModal } from "../components/modals";
 import '../theme/grid.css';
+import ReportError from "../components/reportButton";
 
 const siteTitle = "Domain Tools";
 
@@ -262,6 +263,9 @@ export default function DomainToolsHome({ online }: PageProps) {
 										hideFooter
 										autoHeight
 									/>
+								</Box>
+								<Box my={2}>
+									<ReportError title={`Failed ${currentProtocol} on URL: ${currentURL}`} template="report_dns" />
 								</Box>
 							</Box>
 							: